Rend l'import PFR paresseux dans le plugin cantera (#933)
## Changements

- **Lazy import** : n'est plus importé au démarrage du plugin. Chargé
seulement via lors de l'accès à .
- **Wrapper ** utilisé dans pour éviter le chargement de lors de la
découverte du plugin.
- **Tests de chargement paresseux** : vérifie que l'import du package,
l'accès à , et ne chargent pas prématurément.
- **Correction de tests** : évite la contamination de  entre les tests.

Voir aussi : modifications accumulées sur les workflows CI, le plugin
NMR, , et .
