+++ plugins/plugin-template/.github/workflows/publish_plugin.yml
