+++ plugins/spectrochempy-cantera/tests/test_cantera.py
@@ -1,31 +1,23 @@
 # ruff: noqa: S101  # assert allowed in tests
 
-"""Tests for spectrochempy-cantera plugin."""
+"""Tests for spectrochempy-cantera plugin — PFR scope only."""
 
 from __future__ import annotations
 
-from pathlib import Path
+import sys
 
 import pytest
-from spectrochempy_cantera import PFR
 from spectrochempy_cantera import CanteraPlugin
-from spectrochempy_cantera import equilibrium_composition
-from spectrochempy_cantera import reactor_profile
-from spectrochempy_cantera import thermo_properties
 
 from spectrochempy.api.plugins import PluginCapability
 from spectrochempy.api.plugins import PluginState
 from spectrochempy.api.plugins import check_plugin_compatibility
 from spectrochempy.testing.plugins import PluginTestHarness
 
-_HERE = Path(__file__).parent
-
-# Skip all tests if cantera is not installed
 try:
     import cantera as ct
 
     HAS_CANTERA = True
-    # Check if a default mechanism is available
     ct.Solution("gri30.yaml")
     HAS_GRI30 = True
 except Exception:
@@ -34,18 +26,58 @@ except Exception:
 
 
 # ------------------------------------------------------------------
+# Lazy import behaviour
+# ------------------------------------------------------------------
+
+
+def _clean_cantera_modules():
+    for mod in list(sys.modules):
+        if "spectrochempy_cantera" in mod:
+            del sys.modules[mod]
+
+
+def test_import_does_not_load__pfr():
+    """Importing the plugin package does NOT load ``_pfr`` immediately."""
+    _clean_cantera_modules()
+    import spectrochempy_cantera  # noqa: PLC0415,F401 — import itself is the test action
+
+    assert "spectrochempy_cantera._pfr" not in sys.modules
+
+
+def test_accessing_pfr_loads__pfr():
+    """Accessing ``spectrochempy_cantera.PFR`` triggers the lazy import."""
+    _clean_cantera_modules()
+    import spectrochempy_cantera  # noqa: PLC0415,F401
+
+    assert "spectrochempy_cantera._pfr" not in sys.modules
+    _ = spectrochempy_cantera.PFR
+    assert "spectrochempy_cantera._pfr" in sys.modules
+
+
+def test_registration_does_not_load__pfr():
+    """Calling ``register_simulations()`` does not import ``_pfr``."""
+    _clean_cantera_modules()
+    import spectrochempy_cantera  # noqa: PLC0415,F401 — fresh import so _pfr is not loaded
+
+    plugin = spectrochempy_cantera.CanteraPlugin()
+    entries = plugin.register_simulations()
+    assert "PFR" in [e["name"] for e in entries]
+    assert "spectrochempy_cantera._pfr" not in sys.modules
+
+
+# ------------------------------------------------------------------
 # Plugin metadata and compatibility
 # ------------------------------------------------------------------
 
 
 def test_plugin_metadata():
-    """Plugin has required metadata fields."""
+    """Plugin has required metadata fields (PFR scope only)."""
     plugin = CanteraPlugin()
     assert plugin.name == "cantera"
     assert plugin.version == "0.1.0"
     assert plugin.description
     assert PluginCapability.SIMULATION in plugin.capabilities
-    assert PluginCapability.ANALYSIS in plugin.capabilities
+    assert PluginCapability.ANALYSIS not in plugin.capabilities
     assert PluginCapability.ACCESSOR not in plugin.capabilities
 
 
@@ -53,7 +85,6 @@ def test_plugin_compatibility():
     """Plugin passes full compatibility check."""
     plugin = CanteraPlugin()
     issues = check_plugin_compatibility(plugin)
-    # should report cantera as missing if not installed
     if not HAS_CANTERA:
         assert any("cantera" in issue for issue in issues)
     else:
@@ -66,7 +97,7 @@ def test_plugin_compatibility():
 
 
 def test_registration_with_cantera():
-    """Plugin registers simulations and analyses when cantera is available."""
+    """Plugin registers PFR simulation when cantera is available."""
     if not HAS_CANTERA:
         pytest.skip("cantera not installed")
 
@@ -74,19 +105,14 @@ def test_registration_with_cantera():
     harness.register(CanteraPlugin())
 
     simulations = harness.registry.extensions.list_category("simulation")
-    assert "equilibrium" in simulations
-    assert "reactor_profile" in simulations
-    assert "flame_speed" in simulations
+    assert "PFR" in simulations
 
     analyses = harness.registry.extensions.list_category("analysis")
-    assert "thermo_properties" in analyses
-    assert "kinetic_sensitivity" in analyses
-    assert "spectral_coupling" in analyses
+    assert len(analyses) == 0
 
 
 def test_registration_without_cantera():
     """Plugin is marked FAILED when cantera is missing."""
-    # Simulate by checking the requires check
     harness = PluginTestHarness()
     harness.register(CanteraPlugin())
 
@@ -106,108 +132,41 @@ def test_capability_query():
 
     results = harness.registry.get_by_capability(PluginCapability.SIMULATION)
     names = [r["name"] for r in results]
-    assert "equilibrium" in names
+    assert "PFR" in names
 
 
-def test_package_namespace_exposes_simulations():
-    """Cantera simulations are package-level plugin APIs."""
+# ------------------------------------------------------------------
+# PFR namespace access
+# ------------------------------------------------------------------
+
+
+def test_package_namespace_exposes_pfr():
+    """scp.cantera.PFR is accessible and callable when plugin registered."""
     if not HAS_CANTERA:
         pytest.skip("cantera not installed")
 
-    import spectrochempy as scp
+    import spectrochempy as scp  # noqa: PLC0415
 
     if not scp.plugin_manager.has_plugin("cantera"):
         scp.plugin_manager.register(CanteraPlugin())
 
-    assert scp.cantera.equilibrium is equilibrium_composition
-    assert scp.cantera.reactor_profile is reactor_profile
+    pfr_entry = scp.cantera.PFR
+    assert callable(pfr_entry)
 
 
 def test_cantera_does_not_register_dataset_accessor():
-    """Cantera has no dataset accessor until dataset semantics are explicit."""
-    import spectrochempy as scp
+    """Cantera has no dataset accessor."""
+    import spectrochempy as scp  # noqa: PLC0415
 
     if HAS_CANTERA and not scp.plugin_manager.has_plugin("cantera"):
         scp.plugin_manager.register(CanteraPlugin())
 
     dataset = scp.NDDataset([300.0, 310.0])
     assert not hasattr(dataset, "cantera")
-    assert not hasattr(dataset, "cantera_equilibrium")
 
 
 # ------------------------------------------------------------------
-# Simulation functions
-# ------------------------------------------------------------------
-
-
-@pytest.mark.skipif(not HAS_GRI30, reason="gri30.yaml not available via cantera")
-def test_equilibrium_composition():
-    """Equilibrium composition returns expected structure."""
-    result = equilibrium_composition(
-        mechanism="gri30.yaml",
-        temperature=1200.0,
-        pressure=101325.0,
-        reactants={"CH4": 1.0, "O2": 2.0},
-    )
-
-    assert "species" in result
-    assert "mole_fractions" in result
-    assert "temperature" in result
-    assert result["temperature"] == 1200.0
-    assert result["n_species"] > 0
-    assert result["mole_fractions"].data.ndim >= 1
-
-
-@pytest.mark.skipif(not HAS_GRI30, reason="gri30.yaml not available via cantera")
-def test_thermo_properties():
-    """Thermodynamic properties return expected structure."""
-    result = thermo_properties(
-        mechanism="gri30.yaml",
-        species="CH4",
-        T_min=300.0,
-        T_max=1500.0,
-        n_points=10,
-    )
-
-    assert "temperature" in result
-    assert "cp" in result
-    assert "enthalpy" in result
-    assert "entropy" in result
-    assert "gibbs" in result
-    assert len(result["temperature"].data) == 10
-
-
-@pytest.mark.skipif(not HAS_GRI30, reason="gri30.yaml not available via cantera")
-def test_reactor_profile():
-    """Reactor profile returns expected structure."""
-    result = reactor_profile(
-        mechanism="gri30.yaml",
-        initial_conditions={"T": 1500.0, "P": 101325.0},
-        residence_time=0.01,
-        n_points=10,
-    )
-
-    assert "time" in result
-    assert "temperature" in result
-    assert "species" in result
-    assert "mole_fractions" in result
-    assert len(result["time"]) == 10
-
-
-@pytest.mark.skipif(not HAS_GRI30, reason="gri30.yaml not available via cantera")
-def test_equilibrium_different_reactants():
-    """Equilibrium with different reactant compositions."""
-    result = equilibrium_composition(
-        mechanism="gri30.yaml",
-        temperature=1000.0,
-        pressure=101325.0,
-        reactants={"H2": 2.0, "O2": 1.0},
-    )
-    assert result["n_species"] > 0
-
-
-# ------------------------------------------------------------------
-# PFR tests
+# PFR import and basic checks
 # ------------------------------------------------------------------
 
 
@@ -219,7 +178,7 @@ def test_cantera_is_not_available(monkeypatch):
     assert not _cantera_is_not_available()
 
     with monkeypatch.context() as m:
-        import spectrochempy_cantera._pfr as _pfr  # noqa: PLC0415, PLC0415
+        import spectrochempy_cantera._pfr as _pfr  # noqa: PLC0415
 
         m.setattr(_pfr, "ct", None)
         assert _cantera_is_not_available()
@@ -229,8 +188,10 @@ def test_cantera_is_not_available(monkeypatch):
 
 @pytest.mark.skipif(not HAS_CANTERA, reason="cantera not installed")
 def test_pfr_import():
-    """PFR class can be imported from the cantera plugin."""
-    assert PFR.__name__ == "PFR"
+    """PFR class accessible via ``from spectrochempy_cantera import PFR``."""
+    from spectrochempy_cantera import PFR as _PFR  # noqa: PLC0415,N811
+
+    assert _PFR.__name__ == "PFR"
 
 
 def test_pfr_construction_skipped_due_to_cantera_api():
