+++ plugins/spectrochempy-nmr/pyproject.toml
