+++ plugins/spectrochempy-nmr/tests/test_plugin.py
