+++ src/spectrochempy/plotting/stylesheets/mycustom.mplstyle
